moved adapters tests to individual files
